refactor stuff around
