+++ kensu/airflow/operators/__init__.py
@@ -1,0 +1,3 @@
+from airflow.operators import *
+
+from kensu.airflow.operators import bash, python
