+++ kensu/airflow/providers/google/cloud/operators/__init__.py
@@ -1,0 +1,3 @@
+from airflow.providers.google.cloud.operators import *
+
+from kensu.airflow.providers.google.cloud.operators import bigquery
