+++ kensu/airflow/providers/google/cloud/transfers/__init__.py
@@ -1,0 +1,3 @@
+from airflow.providers.google.cloud.transfers import *
+
+from kensu.airflow.providers.google.cloud.transfers import gcs_to_gcs
